Add Khronos OpenCL loader bazel configuration
ComputeCpp depends on the OpenCL library to communicate with hardware.
Previously, the OpenCL symbols were being resolved dynamically, which
worked well for the Tensorflow wheel, but doesn't for the tests. This
commit adds a minimal bazel configuration for the OpenCL ICD loader
from Khronos and the OpenCL headers, then sets up the dependencies,
such that ComputeCpp depends on the loader, which in turn depends on
the headers. This is more robust as it means that the build depends
less on how the building system is configured.