Improve OpenCL bazel configuration
OpenCL is now a cc_library, meaning that all the kernel tests are linked
against its bits statically. It also uses the exports map from the ICD
repo to indicate versions of the files within it, so should be
compatible with ICD loaders from other vendors. The headers are no
longer an inc_library, as bazel authors indicate that this is not a
useful rule, and library seems to suffice for this case. As of now, the
bazel config will build all tests and the pip package!